Updated the map to to use geolocation to center on the vistor's location. This means not using the embedded map we were previously loading with an iframe.
This has a big side benefit of giving us easy access to all the JavaScript Google Maps API. The downside is that now I need to add the resource locations to the map.

Added an import function to the website. This is extremely basic and needs improvement but it will get us going for local development.
It consists of a html form for uploading a csv file which is stored in Google DataStore by Go code running on App Engine.

I also played around some with css/js on the legend on the map page. Eventually that could be used for filtering the kind of locations shown.

I still need to write the api piece to serve the data now in DataStore.