This commit reorganizes and renames the project some. Also it adds the beginning of a RESTful api. Also it reworks the error pattern to follow examples from the Golang blog and Google Cloud.
Renamed files to remove the "rap" prefix and instead just name the file for what it is with app as the base file.

Moved the code for displaying templated html pages into its own file.

Changed import over from a looping DB update to a batch update which resulted in immediate and obvious performance improvements.

Stubbed out the api's http methods and built the first version of the GET method for it. Also put in some comments on future direction.

At this point there are only a couple tasks remaining for a demoable first version:

* caching on the api
* serve geojson from the api to enable maps integration
* client side js to call the api
* unit tests for templating, import, and api (could require some code reorg)
* integration tests for api (import won't get integration tests until it has insert OR update functionality)